WIP on user, WS and kad handling.
