+++ index.html
@@ -65,6 +65,7 @@ Copyright (C) 2016 The Streembit software development team
     <script type="text/javascript" src="assets/js/streembit/errorhandler.js"></script>
 
     <script type="text/javascript" src="assets/js/streembit/bootclient.js"></script>
+    <script type="text/javascript" src="assets/js/streembit/user.js"></script>
     <script type="text/javascript" src="assets/js/streembit/contacts.js"></script>
     <script type="text/javascript" src="assets/js/streembit/peercomm.js"></script>
